Cambio del nombre del directorio.
