+++ src/main/java/com/trucks_logistics/Trucks/Logistics/auth/IUserService.java
@@ -1,0 +1,5 @@
+package com.trucks_logistics.Trucks.Logistics.auth;
+
+public interface IUserService {
+
+}
