+++ src/main/java/com/trucks_logistics/Trucks/Logistics/auth/User.java
@@ -1,4 +1,4 @@
-package com.trucks_logistics.Trucks.Logistics.users;
+package com.trucks_logistics.Trucks.Logistics.auth;
 
 import java.time.LocalDateTime;
