+++ src/main/java/com/trucks_logistics/Trucks/Logistics/auth/UserController.java
@@ -1,4 +1,4 @@
-package com.trucks_logistics.Trucks.Logistics.users;
+package com.trucks_logistics.Trucks.Logistics.auth;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
