+++ src/main/java/com/trucks_logistics/Trucks/Logistics/auth/UserLoginRequest.java
@@ -1,0 +1,4 @@
+package com.trucks_logistics.Trucks.Logistics.auth;
+
+public record UserLoginRequest() {
+}
