+++ src/main/java/com/trucks_logistics/Trucks/Logistics/auth/UserMapper.java
@@ -1,0 +1,5 @@
+package com.trucks_logistics.Trucks.Logistics.auth;
+
+public class UserMapper {
+
+}
