+++ src/main/java/com/trucks_logistics/Trucks/Logistics/auth/UserService.java
@@ -1,4 +1,4 @@
-package com.trucks_logistics.Trucks.Logistics.users;
+package com.trucks_logistics.Trucks.Logistics.auth;
 
 import org.springframework.stereotype.Service;
