--- src/main/java/com/trucks_logistics/Trucks/Logistics/users/IUserService.java
@@ -1,5 +1,0 @@
-package com.trucks_logistics.Trucks.Logistics.users;
-
-public interface IUserService {
-
-}
