--- src/main/java/com/trucks_logistics/Trucks/Logistics/users/UserLoginRequest.java
@@ -1,4 +1,0 @@
-package com.trucks_logistics.Trucks.Logistics.users;
-
-public record UserLoginRequest() {
-}
