--- src/main/java/com/trucks_logistics/Trucks/Logistics/users/UserRegisterRequest.java
@@ -1,5 +1,0 @@
-package com.trucks_logistics.Trucks.Logistics.users;
-
-public record UserRegisterRequest() {
-
-}
